refactor: move json schemas to config
